Added array textures + example test and more
+ Added support for sRGB texture formats and options when reading image from disk.
+ Added support for array textures - OpenGL implementation complete.
+ Minor improvements and changes; index buffer now takes const uint32_t* as index data to allow constexpr to work.

+++ Quark/src/Platform/OpenGL/OpenGLTextureArray.h
@@ -21,7 +21,7 @@ namespace Quark {
 			return m_RendererID == ((OpenGLTexture2DArray&)other).m_RendererID;
 		}
 
-		virtual void SetData(void* data, uint32_t size, uint32_t layer) override;
+		virtual void SetData(void* data, size_t size, uint32_t layer) override;
 
 		virtual void Attach(uint32_t textureSlot = 0) const override;
 		virtual void Detach() const override;
